src/requirements.txt
boto3==1.26.92
dotmap==1.3.30
jsonschema==4.17.3
requests==2.28.2
simple-salesforce==1.12.3
python-dotenv
asyncio==3.4.3